[change]ingredient/rates migratefile

+++ db/migrate/20200831063500_create_rates.rb
@@ -3,7 +3,7 @@ class CreateRates < ActiveRecord::Migration[5.2]
     create_table :rates do |t|
       t.string :end_user_id, null: false
       t.string :cocktail_id, null: false
-      t.float :rate, null: false
+      t.integer :rate, null: false
 
       t.timestamps
     end
